Remove $PATH from --config and build dumps.
While it is useful to be able to see the user's path in bug reports, it is
perhaps slightly too intrusive to post this without the user's permission.

A path can have usernames or other project sensitive information, and several
Homebrew users were editing their bug reports to omit this information.

`brew doctor` will still report on the path issues that we typically care
about, so dropping automatic posting of PATH.