Don't version VCS checkout directories.
While tarballs are versioned to prevent conflicts when newer ones are downloaded,
we shouldn't be versioning the name of the VCS checkout when a formula uses
"HEAD" or a stable commit.

Append the name of the vcs system to the checkout path, in case a project changes
vcs providers, to prevent future conflicts.