Don't print error message when dev_tools_path fails to find tools
This message is either (a) never printed because the tools are
installed, or (b) printed every time this method is called.

See Homebrew/homebrew#16119.

This code shouldn't be responsible for UX anyway, so just remove it and
let other parts of Homebrew be responsible for failing or declaring that
they need the developer tools.