Added MacHg
MacHg is a free open source OSX GUI for the distributed revision
control system Mercurial. This allows you in a nice graphical way to
manage a collection of files, to add things to the collection, to save
a snapshot of the collection, to restore the collection to an earlier
state and in general to work with the files.