Update snes9x license from unknown to other
Complicated.  Source is available, but not 'open' for the purposes of reusing.  snes9x is gratis for non-commercial purposes.  See the source for the full license but the relevant info:
Snes9x is freeware for PERSONAL USE only. Commercial users should
seek permission of the copyright holders first. Commercial use includes
charging money for Snes9x or software derived from Snes9x.