Update spotifree license from unknown to oss
This one is a little confusing.  There is a license file (https://github.com/ArtemGordinsky/Spotifree/blob/master/Spotifree/License.txt) and it's most, but not all, of the MIT license.
There's a past issue (https://github.com/ArtemGordinsky/Spotifree/issues/19) in which someone claims it's BSD 2-clause, but I see no evidence for that and it's a confusing issue anyway.