Remove mpv.rb which duplicates core formula mpv (#20483)
First, there's the duplication issue (see #15603).

Secondly and more importantly, there has been no official OS X build for
the past three releases (and no sign for its resurgence in the future),
so the cask is stuck on the now-roughly-half-a-year-old 0.14.0 while mpv
as well as its core formula has moved on to 0.17.0. Keeping an
increasingly outdated version in cask won't benefit anyone (one should
assume that a user capable of installing Cask in the first place should
be capable of running `brew install mpv`).

Also see discussion and pointers in #17447 for why mpv is actually a
command line tool.