Codeship Jet, remove Docker dependency (#25226)
With Docker 4 Mac there are other ways besides Homebrew to install Docker on macOS and they are not always compatible (e.g. `brew install docker` fails to link the binary if Docker for Mac is already installed).

Remove the dependency for now, as the documentation at http://documentation.codeship.com/docker/installation/#prerequisites already mentions Docker as a prerequisite.