Update monero-wallet to 0.13.0.3 (#54082)
Their github shows it as a pre-release but the [website shows this as the current release version](https://getmonero.org/downloads/#mac).  This is important because there was a [network upgrade October 18th](https://www.reddit.com/r/Monero/comments/9lcdme/), this older version will show an error and not properly connect to the network.