miniconda - use a stable install path (#66544)
* miniconda - use a stable install path

Conda updates itself (as noted by the use of `auto_updates true`) so as noted by @janosh by installing into `staged_path` we end up using a path with the version number of the initial install of miniconda. So when we update conda with itself using:

```
$ conda update conda
```

We'll end up with a version that's different then the install path indicates which could cause confusion:

```
$ conda info --base
/usr/local/Caskroom/miniconda/4.6.14/miniconda3

$ conda --version
conda 4.7.10
```

This change uses `#{caskroom_path}/base` as the install path which gives us a nicer looking path for the base Conda env:

```
$ conda env list
base                  *  /usr/local/Caskroom/miniconda/base
```

And it will be stable across Conda updates:

```
$ conda info

     active environment : base
    active env location : /usr/local/Caskroom/miniconda/base
            shell level : 1
       user config file : /Users/xyu/.condarc
 populated config files :
          conda version : 4.7.10
    conda-build version : not installed
         python version : 3.7.3.final.0
       virtual packages :
       base environment : /usr/local/Caskroom/miniconda/base  (writable)
           channel URLs : https://repo.anaconda.com/pkgs/main/osx-64
                          https://repo.anaconda.com/pkgs/main/noarch
                          https://repo.anaconda.com/pkgs/r/osx-64
                          https://repo.anaconda.com/pkgs/r/noarch
          package cache : /usr/local/Caskroom/miniconda/base/pkgs
                          /Users/xyu/.conda/pkgs
       envs directories : /usr/local/Caskroom/miniconda/base/envs
                          /Users/xyu/.conda/envs
               platform : osx-64
             user-agent : conda/4.7.10 requests/2.22.0 CPython/3.7.3 Darwin/18.6.0 OSX/10.14.5
                UID:GID : 501:20
             netrc file : None
           offline mode : False
```

* Add caveat about needing conda init to finish setup

In order to use Conda `conda init` must first be run in the user's interactive login shell. Add a caveat to inform users of this after install, this resolves #66490.