Cask ireadfast: 32-bit macOS dependency add depends_on (#118911)
* 32-bit macOS dependency add depends_on

Mojave was the last version of macOS to support 32-bit applications. Unfortunately, this application doesn’t have a 64-bit version for either Intel or Apple Silicon.
So I have added a dependency for Mojave or below.

"Starting with macOS Catalina, 32-bit apps are no longer compatible with macOS.” - Apple

* Update ireadfast.rb

* 32-bit macOS dependency add depends_on

Mojave was the last version of macOS to support 32-bit applications. Unfortunately, this application doesn’t have a 64-bit version for either Intel or Apple Silicon.
So I have added a dependency for Mojave or below.

"Starting with macOS Catalina, 32-bit apps are no longer compatible with macOS.” - Apple

From Discussion 2935