Update Tailscale cask to use .pkg installer
For the past few months, Tailscale has been distributed on its official website using a .pkg installer. Using the .pkg installer is preferred over the .zip file that Homebrew is currently using as source for the application, because the installer includes a preinstall script that disables the VPN tunnel before performing the installation. This reduces the chance of conflicts between versions when performing upgrades.

This PR updates the cask to use the .pkg installer, fixes a capitalization typo in the desc, adds new paths to the zap stanza, and adds two caveats relevant to Tailscale: kext and license.