utm@beta: add livecheck
livecheck checks the Git repository for `utm@beta` using the `Git`
strategy by default but this will be skipped as versioned if/when a
recent `brew` change is merged. This adds a `livecheck` block that
uses the `GithubReleases` strategy to check recent releases and this
`strategy` block includes releases marked as "pre-release".