calmly-writer: update livecheck
This moves the regex out of the `strategy` block and into a `#regex`
call. The `strategy` block also replicates default `HeaderMatch`
logic, so we can simply omit it. Besides that, this tweaks the regex
a bit, in hopes of avoiding potential breakage in the future.

[For what it's worth, the regex is only necessary because otherwise
the `64` in the `-arm64` suffix will be matched by the default
strategy regex.]