transmission@nightly: update version, livecheck
File names for nightly `transmission` releases only include a hash
but we also need the build number to be able to use a fully-versioned
URL. Using a `lastSuccessfulBuild` dmg URL would result in a 404 (Not
Found) response when the next build occurs, as the expected file
won't be available anymore. With that in mind, we need to use a
build-specific dmg URL.

This updates the cask to include the build number in the `version`
and to capture that information in the `livecheck` block regex.
Unfortunately the dmg URL on the `lastSuccessfulBuild` page doesn't
use a build number but we can identify the build number from other
areas of the page.