mediainfo 24.11
This updates `mediainfo` to the latest version on the first-party
download page (https://mediaarea.net/en/MediaInfo/Download/Mac_OS),
24.11. The latest version in the Sparkle feed is 24.10 but the files
for that version are no longer available (see the directory listing
page: https://mediaarea.net/download/binary/mediainfo-gui/).

With that in mind, this updates the `livecheck` block to check the
first-party download page instead of the Sparkle feed. We typically
prefer to align with any in-app updater but not if it contains
outdated versions and/or the latest version isn't available anymore.