opencat: update livecheck
The existing `livecheck` block for `opencat` works as expected but the
`strategy` block uses `split(".").length` to determine the number of
version parts. I recently used `count(".")` in a similar situation,
which accomplishes the same goal but is slightly faster and without
the memory allocation of the `#split`/`#length` approach.

This updates the `strategy` block to use `#count` and reworks the
order of the ternary operator values to prioritize the normal case
(rather than the problematic one).