bambu-studio: update livecheck, version
The `livecheck` block for `bambu-studio` checks the homepage, which
links to the cask `url` dmg but also has a link for a pre-release
(beta) dmg. Both use a stable version format, so the `livecheck`
block is returning the unstable version as newest. The previous
commit addressed this by updating the regex to match the inner text
of the link as a way of avoiding the beta version. That makes the
regex more explicit/fragile and we can handle this in a better way.

This updates the `livecheck` block to use the `GithubLatest` strategy,
matching the version parts (version, suffix, tag version) from the
download URL. Upstream marks unstable (e.g., beta) versions as
pre-release on GitHub, so this will return the latest stable version.

Besides that, this ensures that the `strategy` block only returns
three `version` parts when the filename version differs from the tag
version. The existing `strategy` block attempted to do that but it
was comparing the suffix and tag version (matches 2 and 0) instead of
the file name version and tag version (matches 1 and 0), so the
current version contains three version parts when we only need the
first two. This commit also updates the `version` accordingly.