pgadmin4: update livecheck
The existing `livecheck` block for `pgadmin4` checks the version
directories on the directory listing page related to the cask `url`
source. However, upstream has created a `v9.1` directory that doesn't
contain any release assets yet, so this check is erroneously returning
9.1 instead of 9.0.

The first-party download page correctly links to 9.0 as the newest
version, so this updates the `livecheck` block to check that instead.