readyapi: update livecheck
The existing `livecheck` block for `readyapi` is returning an `Unable
to get versions` error, as the text on the page has changed and the
expected text is no longer present. The latest version is still on
the page but trying to match that text may prove to be similarly
unreliable over time.

This updates the `livecheck` block to check the Version History page,
which lists the versions in a more predictable manner.