batfi: update livecheck
The existing `livecheck` block for `batfi` checks the Sparkle appcast
and it's currently returning 2.3.3 as newest but that version is on
the beta channel and the file URL is `BatFi-beta.zip` instead of
`BatFi-latest.zip` (which is 2.3.2). This updates the `livecheck`
block to add a `strategy` block that finds the newest release on the
main channel.