astah-professional: update livecheck
The existing `livecheck` block for `astah-professional` is giving an
`Unable to get versions` error, as the URL now redirects to the
Release Notes page. This updates the `livecheck` block to check the
first-party download page, which contains a URL for the newest dmg
file (albeit, in an `onclick` attribute instead of `href`).