maintenance 3.3.1
This updates `maintenance` to the latest version for Sequoia, 3.3.1,
and updates the `livecheck` block setup. The homepage states,
"Titanium Software has discontinued developing the old versions and
will no longer update them", so I've added `skip` `livecheck` blocks
to the older versions and moved the existing `livecheck` block into
the newest on_system block.

I updated the `livecheck` block to use the approach from the `calhash`
cask, which checks the versions on the homepage. The Sequoia plist
file for `maintenance` lists 3.2.9 as the newest version but the
homepage says it's 3.3.1, so we can't rely on the plist file being
kept up to date.

Besides that, I tweaked the regex and replaced the interpolated
`MacOS.version` with a hardcoded version, so this will work
predictably regardless of the execution environment.