gns3: update livecheck
The existing `livecheck` block for `gns3` uses the `GithubLatest`
strategy but upstream creates releases for multiple major versions
(e.g., 3.0.4, 2.2.54) and the "latest" release isn't guaranteed to be
the highest version. Unfortunately, the download page on www.gns3.com
simply uses the "latest" GitHub release version, so it's also
affected.

At the moment, 2.2.54 is the "latest" release but the highest version
is 3.0.4 (the previous release). This addresses the issue by updating
the `livecheck` block to use the `GithubReleases` strategy.