swiftformat-for-xcode 0.56.1
`swiftformat-for-xcode` is in the autobump list but the 0.56.1 update
is failing because the filename unpredictably alternates between
`SwiftFormat.for.Xcode.zip` and `SwiftFormat.for.Xcode.app.zip`, so
updating breaks with each switch. In this case, the 0.56.0 release
did not include `.app` in the filename and 0.56.1 does. This
temporarily resolves the issue by manually updating the formula.