script-debugger: update livecheck
The existing `livecheck` block for `script-debugger` is returning an
`Unable to get versions` error, as the HTML structure of the download
page has changed and the regex no longer matches. This updates the
regex to replace the leading `action=` with `href=`, as the URLs are
now found in `href` attributes. Besides that, this also tweaks the
regex to align it more with prevailing standards and to make the
trailing suffix optional (as older versions don't have this suffix
and we don't want to miss future versions if it's omitted again).