retroshare: update livecheck
Code scanning has flagged the `livecheck` block regex for `retroshare`
as an inefficient regular expression, specifically the `\d+` part of
the `(\d+(?:\.\d+)+[a-z]?)+` group. We use `\d+(?:\.\d+)+[a-z]?` in
other regexes without any issue, so this is either due to the regex
having a preceding `(\d+(?:\.\d+)+)` group or because the
aforementioned group is allowed to repeat and greedily (i.e., it has
`+` on the end). From what I'm seeing in previous cask `version`
values, it doesn't seem like that part of the regex needs to repeat
(it works for the current filenames, at least), so let's make it
non-repeating and see if that resolves the code scanning alert.