sqlcl 25.2.1.195.1751
The autobump workflow isn't able to update `sqlcl` to the latest
version because the existing homepage now returns a 404 (Not Found)
response. This updates the homepage to the current URL while updating
to the latest version, 25.2.1.195.1751.

Besides that, this also updates the `livecheck` block regex to match
the version from the version in the zip file URL, which is more in
line with typical patterns.