picoscope@beta 7.2.3.6884
`picoscope@beta` is autobumped but the workflow is unable to update to
the latest version because the `livecheck` block regex isn't able to
match the current file URL on the upstream download page. This updates
the regex to be able to match the current file URL and updates the
cask to the latest version.