pronotes: update livecheck
The HTTP response for the `pronotes` appcast contains a
`Content-Encoding: aws-chunked` header and this causes curl to return
an "Unrecognized content encoding type" error when the `--compressed`
option is used. livecheck uses `--compressed` when fetching content,
so this check will predictably fail until we can selectively disable
compression in a `livecheck` block.

For the time being, this comments out the existing `url`/`regex` and
adds a `skip` call with an explanatory comment.