atv-remote: add livecheck
By default, livecheck checks the Git tags for `atv-remote` but
upstream marks some releases that use a stable version format (e.g.,
v1.2.3) as pre-release on GitHub. The newest release, v1.4.1, is
pre-release and the description contains a "this should probably be a
release version" comment, which seems to suggest that releases marked
as pre-release aren't considered stable versions.

This adds a `livecheck` block that uses the `GithubLatest` strategy,
so livecheck will return the newest non-pre-release version (1.3.7).