xcodes-app: add livecheck
By default, livecheck checks the Git tags for `xcodes-app` and
returns 3.0.0b32 as the newest version but the related GitHub release
is titled "v3.0.0 Beta 1" and marked as pre-release. This adds a
`livecheck` block that uses the `GithubLatest` strategy, which
correctly returns 2.4.2b31 as the newest stable version (though we
need a regex to handle the version format).