qcad 3.32.4
`qcad` is autobumped but the workflow failed to update to 3.32.4
because the upstream download page also contains links for QCAD
Snapshot that use the same file name format as QCAD Professional
(what the cask is using) but have a higher version and those URLs
just happened to return a 404 (Not Found) response. This updates the
cask and modifies the `livecheck` block to only use the version from
the first matching link on the download page, which should be for
QCAD Professional.