pcsx2: update livecheck, add app target
This updates the `livecheck` block for `pcsx2` to use the same API
endpoint that the app uses to check for updates. The `GithubLatest`
strategy works but older releases don't include a macOS asset and we
try to align checks with the in-app updater when possible.

Besides that, this adds an app `target` value to omit the version
from the app name.