zipic 1.8.1
`zipic` is autobumped but the workflow failed to update to 1.8.1
because `brew audit` failed with an "Upstream defined :ventura as the
minimum macOS version but the cask declared a depends_on stanza with
a minimum macOS version of :monterey" error. This updates the version
and modifies the `depends_on macos:` value accordingly.