quo 4.1.12
`openphone` is autobumped but the workflow failed to update to 4.1.12
because the app has been renamed to Quo, so the file name changed as
well. This updates and renames the cask accordingly. For what it's
worth, the appcast URL remains the same in the Quo app (and updating
the domain in the path didn't work), so we will have to update it
if/when it changes in the future.