toinane-colorpicker 2.3.0
`toinane-colorpicker` is autobumped but the workflow failed to update
to version 2.3.0 because the Intel dmg file name doesn't use an `-x64`
suffix in this release. `brew audit` also fails with an "Artifact
defined :monterey as the minimum macOS version but the cask declared
no minimum macOS version" error. This updates the cask accordingly.