geoda 1.22.0.21
`geoda` is autobumped but the workflow failed to update to 1.22.0.21
because livecheck is returning an `Unable to get versions` error. The
latest macOS files for the v1.22.0.21 release are now zip files and
the file name format has changed, so the existing regex wasn't able
to match. This updates the version and adjusts the `livecheck` block
regex to handle both the previous and current file name formats.