msgfiler: downgrade to 4.2.0.41
The 4.3.0 version of `msgfiler` is listed as beta on the first-party
download page but livecheck has been returning those versions as
newest because the dmg file name uses the same format as the stable
version. This downgrades the cask to the newest stable version and
updates the `livecheck` block to only match the stable versions.