espresso 25.10
The `espresso` zip URL changed with the 2.9 release, so this updates
the cask accordingly. The app is now a universal binary, so the
Rosetta caveat is no longer necessary.

Besides that, this updates the `livecheck` block URL to align with the
download URL on the homepage (it's basically the same content as the
`/updates/` page) and to make the regex a bit more flexible about
matching the attribute value.