gitfox 4.0.2
`gitfox` is autobumped but the workflow failed to update to version
4.0.1 because the cask URL format doesn't work for this release.
The download links on the "Older Versions" page and the URLs in the
API response from the `livecheck` block URL both use update.gitfox.app
URLs instead of storage.googleapis.com. This updates the version and
modifies the `url` accordingly.