folo 0.7.1
`folo` is autobumped but the workflow failed to update to version
0.7.1 because upstream is now using a `desktop/v1.2.3` tag format and
livecheck was giving an `Unable to get versions` error due to the
regex not matching the new tag format. This updates the version and
modifies the `livecheck` block regex to allow for a `desktop` prefix
(the newest release uses `desktop/` but there are other tags using
`desktop@`, so I've set it to match both).