flowdown 3.8.433
`flowdown` is autobumped but the workflow failed to update to version
3.8.433 because the current zip file name doesn't include a `v` before
the version (upstream unpredictably includes/omits it). Besides that,
`brew audit` also fails with an "Artifact defined :sonoma as the
minimum macOS version but the cask declared a depends_on stanza with
a minimum macOS version of :ventura" error. This updates the version,
url, and `depends_on macos:` value accordingly.