appflowy 0.10.3
`appflowy` is autobumped but the workflow failed to update to version
0.10.3 because the GitHub release asset for the Intel zip file
mistakenly uses a 0.10.2 version in the file name. This works around
the issue by switching the cask to use the dmg files instead, which
doesn't have the same problem for this release.

Besides that, this adds a `depends_on macos:` value to address the
related `brew audit` error ("Artifact defined :big_sur as the minimum
macOS version but the cask declared no minimum macOS version").