dynobase 2.5.1
`dynobase` is autobumped but the workflow failed to update to versions
beyond 2.3.0 because the cask uses a GitHub release asset but the
GitHub repository has been deleted. Upstream now serves files from
download.todesktop.com but they require users to jump through hoops to
download the app. Rather than link to the dmg files directly, they
provide an installer that checks a JSON file that includes the latest
version and dmg file URLs and uses that information to install the
app.

This updates the version and `url` using the information from the JSON
file and updates the `livecheck` block to check this file.