elektron-overbridge: update livecheck
`elektron-overbridge` is autobumped but the livecheck block has been
consistently producing an "Unable to get versions" error for quite
some time. The `livecheck` block fails because the website is behind
Cloudflare and the protections prevent us from fetching pages outside
of a browser context (i.e., this fails regardless of user agent, IP
address, etc.). This updates the `livecheck` block to skip and adds a
comment with the URL for manual checking.