insta360-studio 5.8.0
`insta360-studio` is autobumped but the workflow has been unable to
check for new versions as livecheck is returning an "Unable to get
versions" error. This is due to the file name for 5.8.0 including an
underscore between "Insta360" and "Studio", which the regex wasn't
set up to handle. This was an oversight on my part, as I used a
`(?:[._-]|%20)` pattern in other parts of the regex and should have
done the same in this area.

This updates the version and `url` and adjusts the `livecheck` block
regex accordingly.