liquibase-secure: update livecheck
The existing `livecheck` block for `liquibase-secure` is producing an
"Unable to get versions" error because the `href` values for the
version directory links are now absolute rather than relative, so the
regex is unable to match. This addresses the issue by updating the
regex to be able to match both absolute and relative URLs in this
context.