macx-dvd-ripper-pro: update livecheck
The existing `livecheck` block for `macx-dvd-ripper-pro` produces an
"Unable to get versions" error, as the homepage no longer contains
any version information. This updates the `livecheck` block to check
the URL that the app uses to check for updates. The related plist
file provides a `LastestVersion` key that provides the latest version,
so I've also included some fallback login in the `strategy` block to
check for a `LatestVersion` key if upstream fixes this typo.