mysql-shell 9.5.0
`mysql-shell` is autobumped but the upstream website returns a 403
(Forbidden) response unless brew requests uses a "curl/8.7.1" user
agent, so the `livecheck` block predictably returns an "Unable to get
versions" error until we can set the user agent there (I have this
implemented and confirmed it works here but still need to create a
brew PR for this work). In the interim time, this manually updates the
cask to the newest upstream version.