mysqlworkbench 8.0.44
`mysqlworkbench` is autobumped but the workflow failed to update to
version 8.0.44 because the `livecheck` block produces an "Unable to
get versions" error, as brew requests to dev.mysql.com receive a 403
(Forbidden) response unless they use a "curl/8.7.1" user agent. This
feature isn't available in `livecheck` blocks yet but I confirmed it
works with my local implementation (I'll create a PR for user agent
support sometime in the near future). In the interim time, this
manually updates the cask to the newest upstream version.

Besides that, this updates the `livecheck` block to align with the
approach used in the `mysql-shell` cask, which uses query string
parameters to ensure that the HTML includes information for the macOS
artifacts. The page automatically selects the related OS option based
on the user agent in the HTTP request but the Mac option won't be
selected for a generic "curl/8.7.1" user agent, so this approach is
also necessary for this check to work once user agent support is
available.