nomachine 9.2.18_1
`nomachine` is autobumped but it hasn't been updated since version
8.16.1_2 because the `livecheck` block is broken. The download page
requires a cookie or the server will redirect to a login page instead,
which seems to be what's happening with the existing check.

It's possible to check the redirecting download URL from the homepage
but it only redirects to the dmg file for macOS if the request user
agent references macOS. It's not currently possible to explicitly set
the user agent in the `livecheck` block but livecheck contains some
fallback logic where it will retry a failing request with the
`:browser` user agent, so this check will technically work in the
interim time (albeit, less efficiently). This updates the version and
`livecheck` block accordingly.