nordic-nrf-command-line-tools: update livecheck
The existing `livecheck` block for `nordic-nrf-command-line-tools` is
producing an "Unable to get versions" error because the upstream
website is behind Cloudflare and requires users to go through a
turnstile check in a browser to view any pages. This appears to occur
regardless of user agent, IP address, etc., so this updates the
`livecheck` block to skip and adds a comment with the URL to manually
check.