teamspeak-client@beta 6.0.0-beta3.2
`teamspeak-client@beta` is autobumped but the workflow failed to
update to version 6.0.0-beta3.2 because the `livecheck` block is
producing an "Unable to get versions" error. The existing regex is
unable to match the current dmg file names because they include an
arch suffix now that there's also a dmg for ARM. This updates the
version and adjusts the cask accordingly.

Besides that, the app fails Gatekeeper checks, so this also adds a
`disable!` call with the future date we've been using for this
scenario.