unite-phone 2025.10.7
`unite-phone` is autobumped but the workflow has failed to update to
versions beyond 2025.9.2 because the `livecheck` block seems to
produce an "Unable to get versions" error in the autobump environment.
The check worked fine for me locally but the cask `url` uses
`user_agent: :fake`, so the `livecheck` block URL may need the same
treatment to work for autobump. We can't set the user agent in a
`livecheck` block yet (this is forthcoming), so this manually updates
the version in the interim time. Besides that, the app fails
Gatekeeper checks, so this adds a `disable!` call with the future date
we've been using for this scenario.