aionui: update livecheck
The existing `livecheck` block for `aionui` checks the Git tags and is
returning 1.5.0 as newest but there isn't a corresponding release on
GitHub six hours after the tag was created. This updates the
`livecheck` block to use the `GithubLatest` strategy, as there can be
a notable gap between tag and release.