archi: update livecheck
The existing `livecheck` block for `archi` is producing an "Unable to
get versions" error, as the dmg file URLs now reference a `5-7`
release tag on GitHub but the regex doesn't expect this format and
can't match it. This updates the `livecheck` block regex to be able to
match this part of the URL path and adjusts the cask `version`
accordingly (as the dmg files were unreachable).