cmake-app: add macOS dependency
`cmake-app` uses a universal app and the `Info.plist` doesn't specify
a minimum system version, so `brew audit` identifies the requirements
from the binary. In this case, the minimum macOS version is 10.13 or
later for Intel and 11 or later for ARM but the audit simply gives an
"Artifact defined :big_sur as the minimum macOS version but the cask
declared no minimum macOS version" error when run on ARM. We've been
dealing with a similar situation with separate ARM/Intel apps, where
the Intel app can use the implicit macOS dependency but the ARM
version depends on a higher macOS version, so we also have to add a
`depends_on macos:` value for Intel that makes the implicit value
explicit. This is effectively the same issue but with a universal app
instead, so this addresses the issue using on_os blocks (with the
Intel value being added to appease `brew style`).